Added a connection proxy plugin
Introduces a new ioplug plugin for alsa.
This plugin always keeps the opened PCM in a consitent state
for the sound application, despite of bluealsa server or
or device disconnection.
For now, only playback is supported, and only one PCM per application
is allowed.
The sound application is in charge of telling the plugin what device
to use; the bluealsa_proxy_set_remote_device can be opened through
snd_dlopen. Please see the provided README.md for more details.